requirements.txt
altgraph==0.17
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
future==0.18.2
idna==2.10
msedge-selenium-tools==3.141.3
pefile==2019.4.18
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
PyQt5==5.15.2
PyQt5-sip==12.8.1
pywin32-ctypes==0.2.0
requests==2.25.1
selenium==3.141.0
soupsieve==2.1
urllib3==1.26.2
webdriver-manager==3.2.2
